Significant rework of incoming calls processing to fix some aspects of multilink server functionality.
Was:
When incoming call received, physical level calls IfaceOpenNcps(),
which opens bundle. Bundle at his place opens ALL links.
Even that which do not have incoming calls.

Now:
When incoming call received, physical level calls LinkOpen().
When this link connected and authenticated it calls BundJoin()
which (if this is a first link in multilink) calls IfaceOpenNcps().